requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
black==25.1.0
click==8.2.1
Faker==37.4.0
greenlet==3.2.3
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
playwright==1.52.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.1.0
typing_extensions==4.14.0
tzdata==2025.2
